feat: update types from backend, add delrole command

+++ src/commands/role.ts
@@ -13,9 +13,9 @@ const numberOrString = z.string().transform((s) => {
 
 _commandsBase
   .createCommand({
-    trigger: "getrole",
+    trigger: "getroles",
     scope: "private",
-    description: "Get role of userid",
+    description: "Get roles of an user",
     args: [
       {
         key: "username",
@@ -34,7 +34,9 @@ _commandsBase
 
       try {
         const { roles } = await api.tg.permissions.getRoles.query({ userId })
-        await context.reply(fmt(({ b }) => (roles ? [`Roles:`, b`${roles.join(" ")}`] : "This user has no roles")))
+        await context.reply(
+          fmt(({ b }) => (roles?.length ? [`Roles:`, b`${roles.join(" ")}`] : "This user has no roles"))
+        )
       } catch (err) {
         await context.reply(`There was an error: \n${String(err)}`)
       }
@@ -65,20 +67,73 @@ _commandsBase
       }
 
       try {
-        const { roles } = await api.tg.permissions.getRoles.query({ userId })
-        const { error } = await api.tg.permissions.addRole.mutate({ userId, adderId: context.from.id, role: args.role })
+        const { roles, error } = await api.tg.permissions.addRole.mutate({
+          userId,
+          adderId: context.from.id,
+          role: args.role,
+        })
+
         if (error) {
           await context.reply(fmt(({ n }) => n`There was an error: ${error}`))
           return
         }
+
         await context.reply(
           fmt(
-            ({ b, n, u }) => [
-              b`✅ Role set!`,
-              n`${b`Username:`} ${args.username}`,
-              n`${b`Roles:`} ${u`${args.role}`} ${roles ? roles.join(" ") : ""}`,
-            ],
-            { sep: "\n" }
+            ({ b, n }) => [b`✅ Role added!`, n`${b`Username:`} ${args.username}`, n`${b`Updated roles:`} ${roles}`],
+            {
+              sep: "\n",
+            }
+          )
+        )
+        await context.deleteMessage()
+      } catch (err) {
+        await context.reply(`There was an error: \n${String(err)}`)
+      }
+    },
+  })
+  .createCommand({
+    trigger: "delrole",
+    scope: "private",
+    description: "Remove role from an user",
+    args: [
+      {
+        key: "username",
+        type: numberOrString,
+        description: "The username or the user id of the user you want to remove the role from",
+      },
+      { key: "role", type: z.enum<Role[]>(["owner", "president", "direttivo", "hr", "admin"]) },
+    ],
+    permissions: {
+      allowedRoles: ["owner", "direttivo"],
+    },
+    handler: async ({ context, args }) => {
+      const userId: number | null =
+        typeof args.username === "string" ? await getTelegramId(args.username.replaceAll("@", "")) : args.username
+
+      if (userId === null) {
+        await context.reply("Not a valid userId or username not in our cache")
+        return
+      }
+
+      try {
+        const { roles, error } = await api.tg.permissions.removeRole.mutate({
+          userId,
+          removerId: context.from.id,
+          role: args.role,
+        })
+
+        if (error) {
+          await context.reply(fmt(({ n }) => n`There was an error: ${error}`))
+          return
+        }
+
+        await context.reply(
+          fmt(
+            ({ b, n }) => [b`✅ Role removed!`, n`${b`Username:`} ${args.username}`, n`${b`Updated roles:`} ${roles}`],
+            {
+              sep: "\n",
+            }
           )
         )
         await context.deleteMessage()
